Cache Download

By default, all downloads are cached both on the disk and in your computer memory. This means that the second time you download a file during the same session it will be reloaded from your ram. If you stop a session and restart it, the file will be loaded from your disk. If you want to reload the data from its original source, use the bust_cache function.

Usage

cache_download(
  url,
  ...,
  use_memory = TRUE,
  use_disk = TRUE,
  bust_cache = FALSE
)

Arguments

url

The URL to download from.

...

arguments to forward to the vroom::vroom function.

use_memory

Whether to use memory caching.

use_disk

Whether to use disk caching.

bust_cache

Whether to bust (refresh) the cache.

Details

Additionally, options can be used to determine the default caching behavior of tidyiddr.

Examples

cache_download("https://s3.amazonaws.com/quartzdata/datasets/caserates-by-age.csv")
#> # A tibble: 128 × 11
#>    `0 - 4` `5 - 11` `12 - 15` `16 - 17` 18 - 2…¹ 30 - …² 40 - …³ 50 - …⁴ 65 - …⁵
#>      <dbl>    <dbl>     <dbl>     <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1  93.4   171.      151.       155.     141.    1.70e+2 1.54e+2 1.29e+2 108.   
#>  2 105.    147.      218.       279.     328.    3.02e+2 3.00e+2 2.68e+2 207.   
#>  3 202.    102.       92.6      129.     246.    2.72e+2 2.53e+2 2.38e+2 228.   
#>  4   0.266   0.0492    0.0478     0        0.141 1.13e-1 1.51e-1 1.70e-1   0.156
#>  5 149.    170.      170.       172.     209.    2.20e+2 1.98e+2 1.71e+2 147.   
#>  6  53.3    73.9     111.       142.     169.    1.46e+2 1.40e+2 1.25e+2  88.6  
#>  7 112.    230.      176.       175.     169.    2.00e+2 1.81e+2 1.48e+2 111.   
#>  8 159.    299.      348.       360.     282.    2.94e+2 2.59e+2 1.94e+2 153.   
#>  9 764.    853.     1064.      1348.    1959.    1.85e+3 1.65e+3 1.23e+3 820.   
#> 10   0.215   0.218     0.323      0.579    1.43  2.43e+0 3.29e+0 4.04e+0   3.77 
#> # … with 118 more rows, 2 more variables: `75+` <dbl>, week <date>, and
#> #   abbreviated variable names ¹​`18 - 29`, ²​`30 - 39`, ³​`40 - 49`, ⁴​`50 - 64`,
#> #   ⁵​`65 - 74`

options("tidyiddr_use_cache" = TRUE)
cache_download("https://s3.amazonaws.com/quartzdata/datasets/caserates-by-age.csv")
#> [1] "Reading from disk cache"
#> Rows: 128 Columns: 11
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl  (10): 0 - 4, 5 - 11, 12 - 15, 16 - 17, 18 - 29, 30 - 39, 40 - 49, 50 - ...
#> date  (1): week
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 128 × 11
#>    `0 - 4` `5 - 11` `12 - 15` `16 - 17` 18 - 2…¹ 30 - …² 40 - …³ 50 - …⁴ 65 - …⁵
#>      <dbl>    <dbl>     <dbl>     <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1  93.4   171.      151.       155.     141.    1.70e+2 1.54e+2 1.29e+2 108.   
#>  2 105.    147.      218.       279.     328.    3.02e+2 3.00e+2 2.68e+2 207.   
#>  3 202.    102.       92.6      129.     246.    2.72e+2 2.53e+2 2.38e+2 228.   
#>  4   0.266   0.0492    0.0478     0        0.141 1.13e-1 1.51e-1 1.70e-1   0.156
#>  5 149.    170.      170.       172.     209.    2.20e+2 1.98e+2 1.71e+2 147.   
#>  6  53.3    73.9     111.       142.     169.    1.46e+2 1.40e+2 1.25e+2  88.6  
#>  7 112.    230.      176.       175.     169.    2.00e+2 1.81e+2 1.48e+2 111.   
#>  8 159.    299.      348.       360.     282.    2.94e+2 2.59e+2 1.94e+2 153.   
#>  9 764.    853.     1064.      1348.    1959.    1.85e+3 1.65e+3 1.23e+3 820.   
#> 10   0.215   0.218     0.323      0.579    1.43  2.43e+0 3.29e+0 4.04e+0   3.77 
#> # … with 118 more rows, 2 more variables: `75+` <dbl>, week <date>, and
#> #   abbreviated variable names ¹​`18 - 29`, ²​`30 - 39`, ³​`40 - 49`, ⁴​`50 - 64`,
#> #   ⁵​`65 - 74`